+++ CHANGELOG.md
@@ -23,6 +23,8 @@ All notable changes to the "vscode-didact" extension will be documented in this 
 - Added `Insert Didact Badge` completion to show `Powered by Didact` badge with link back to the project in GitHub
 - Added new `vscode` URI processing to handle registering a Didact tutorial via a web link and a new command - `Didact: Process VSCode link from web` to process a copied link
 - Add new `Ctrl/Cmd+Alt+T` shortcut to convert selected text into a `sendNamedTerminalAString` Didact link
+  - Added new settings to customize the link text shown and whether to execute the command in the terminal
+  - Add new `sendNamedTerminalAStringNoLF` command that sends text to the terminal but doesn't execute it
 
 ## 0.3.2